chore: reorganize event type files
