--- apps/web/components/eventtype/EventWorkfowsTab.tsx
@@ -1,1 +1,0 @@
-export { default } from "@calcom/features/ee/workflows/components/EventWorkflowsTab";
