+++ apps/web/test/lib/CheckForEmptyAssignment.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { checkForEmptyAssignment } from "../../lib/checkForEmptyAssignment";
+import { checkForEmptyAssignment } from "@calcom/lib/event-types/utils/checkForEmptyAssignment";
 
 describe("Tests to Check if Event Types have empty Assignment", () => {
   it("should return true if managed event type has no assigned users", () => {
