+++ packages/features/availability/components/SkeletonLoader.tsx
@@ -1,6 +1,5 @@
+import classNames from "@calcom/lib/classNames";
 import { Button, SkeletonText } from "@calcom/ui";
-
-import classNames from "@lib/classNames";
 
 function SkeletonLoader() {
   return (
