+++ packages/features/eventtypes/components/event-type/tabs/ai/EventAITab.tsx
@@ -1,4 +1,4 @@
-import type { EventTypeSetupProps } from "pages/event-types/[type]";
+import type { EventTypeSetupProps } from "@calcom/features/eventtypes/lib/types";
 
 import AIEventController from "./AIEventController";
