+++ packages/features/eventtypes/components/event-type/tabs/instant/EventInstantTab.tsx
@@ -1,5 +1,4 @@
-import type { EventTypeSetupProps } from "pages/event-types/[type]";
-
+import type { EventTypeSetupProps } from "@calcom/features/eventtypes/lib/types";
 import getPaymentAppData from "@calcom/lib/getPaymentAppData";
 
 import InstantEventController from "./InstantEventController";
