+++ packages/lib/event-types/utils/checkForEmptyAssignment.ts
@@ -1,4 +1,7 @@
-import type { EventTypeAssignedUsers, EventTypeHosts } from "~/event-types/views/event-types-single-view";
+import type {
+  EventTypeAssignedUsers,
+  EventTypeHosts,
+} from "@calcom/features/eventtypes/components/event-type/EventType";
 
 // This function checks if EventType requires assignment.
 // returns true: if EventType requires assignment but there is no assignment yet done by the user.
